requirements.txt
aiogram==3.10.0
python-dotenv==1.0.1
asyncpg==0.29.0
alembic==1.13.2
sqlalchemy==2.0.31
redis==5.0.3
aiocron